refactor(observe): fold log forwarding adapters into subpaths (#926)
## Context
Part 4 of the stack. This resolves standalone LogTape/Pino package placement by folding the forwarding adapters into `@ontrails/observe` subpaths.

## What changed
- Moved LogTape and Pino forwarding adapters into observe subpath exports.
- Removed the standalone observe-adapter workspaces and refreshed workspace metadata.
- Updated README/docs/examples and package release intent around the new subpaths.

## Verification
- `bun test packages/observe/src/__tests__/logtape.test.ts packages/observe/src/__tests__/pino.test.ts`
- `bun install --lockfile-only`
- `bun run typecheck`
- `bun run lint`
- `bun run docs:snippets`
- `bun run docs:links`
- `bun run dead-code`
- `bun run publish:check`
- CI is green on this PR.

## Risks / rollout
This removes standalone package surfaces in favor of observe subpaths; release notes and README examples document the replacement import path.
