+++ .changeset/logtape-observe-target.md
@@ -1,6 +1,6 @@
 ---
-"@ontrails/logtape": major
+"@ontrails/observe": major
 ---
 
-Retarget the LogTape adapter to `@ontrails/observe` log sink types and use `createLogtapeSink` as the canonical factory name.
+Retarget the LogTape adapter to the `@ontrails/observe/logtape` subpath and use `createLogtapeSink` as the canonical factory name.
 Refresh the LogTape forwarding migration note for consumers moving off the retired logging package.
