+++ .changeset/observability-migration-docs.md
@@ -1,7 +1,6 @@
 ---
 "@ontrails/observe": patch
 "@ontrails/tracing": patch
-"@ontrails/logtape": patch
 ---
 
-Add migration guidance for the retired `@ontrails/logging` package and align observability README examples around `@ontrails/observe`, `@ontrails/tracing`, and `@ontrails/logtape`.
+Add migration guidance for the retired `@ontrails/logging` package and align observability README examples around `@ontrails/observe`, `@ontrails/tracing`, and `@ontrails/observe/logtape`.
