+++ .changeset/observe-log-forwarding-subpaths.md
@@ -1,0 +1,5 @@
+---
+"@ontrails/observe": major
+---
+
+Fold LogTape and Pino log forwarding into `@ontrails/observe/logtape` and `@ontrails/observe/pino`, and remove the standalone `@ontrails/logtape` and `@ontrails/pino` workspaces.
