+++ .changeset/pre.json
@@ -20,10 +20,8 @@
     "@ontrails/drizzle": "1.0.0-beta.14",
     "@ontrails/hono": "1.0.0-beta.14",
     "@ontrails/vite": "1.0.0-beta.14",
-    "@ontrails/logtape": "1.0.0-beta.14",
     "@ontrails/commander": "1.0.0-beta.15",
     "@ontrails/oxlint-plugin": "1.0.0-beta.15",
-    "@ontrails/pino": "1.0.0-beta.17",
     "@ontrails/adapter-kit": "1.0.0-beta.18",
     "@ontrails/regrade": "1.0.0-beta.18",
     "@ontrails/library": "1.0.0-beta.24",
