+++ .changeset/remove-legacy-logging-package.md
@@ -1,10 +1,9 @@
 ---
 "@ontrails/observe": minor
-"@ontrails/logtape": patch
 "@ontrails/trails": patch
 "@ontrails/testing": patch
 "@ontrails/tracing": patch
 "@ontrails/warden": patch
 ---
 
-Remove the retired `@ontrails/logging` workspace from the prerelease package set. Use `@ontrails/observe` for log and trace sink contracts and `@ontrails/logtape` for LogTape forwarding.
+Remove the retired `@ontrails/logging` workspace from the prerelease package set. Use `@ontrails/observe` for log and trace sink contracts and `@ontrails/observe/logtape` for LogTape forwarding.
