+++ .changeset/trl-722-pino-docs.md
@@ -1,6 +1,5 @@
 ---
 '@ontrails/observe': patch
-'@ontrails/pino': patch
 ---
 
 Document Pino sink usage and publish-readiness checks.
