--- packages/logtape/CHANGELOG.md
@@ -1,142 +1,0 @@
-# @ontrails/logtape
-
-## 1.0.0-beta.39
-
-## 1.0.0-beta.38
-
-## 1.0.0-beta.37
-
-## 1.0.0-beta.36
-
-## 1.0.0-beta.35
-
-## 1.0.0-beta.34
-
-## 1.0.0-beta.33
-
-## 1.0.0-beta.32
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.32
-
-## 1.0.0-beta.31
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.31
-
-## 1.0.0-beta.30
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.30
-
-## 1.0.0-beta.29
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.29
-
-## 1.0.0-beta.28
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.28
-
-## 1.0.0-beta.27
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.27
-
-## 1.0.0-beta.26
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.26
-
-## 1.0.0-beta.25
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.25
-
-## 1.0.0-beta.24
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.24
-
-## 1.0.0-beta.23
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.23
-
-## 1.0.0-beta.22
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.22
-
-## 1.0.0-beta.21
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.21
-
-## 1.0.0-beta.20
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.20
-
-## 1.0.0-beta.19
-
-### Patch Changes
-
-- Updated dependencies [1eb5bdc]
-  - @ontrails/observe@1.0.0-beta.19
-
-## 1.0.0-beta.18
-
-### Patch Changes
-
-- Updated dependencies [bf44972]
-- Updated dependencies [e0ae995]
-  - @ontrails/observe@1.0.0-beta.18
-
-## 1.0.0-beta.17
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.17
-
-## 1.0.0-beta.16
-
-### Major Changes
-
-- 68f70fb: Retarget the LogTape adapter to `@ontrails/observe` log sink types and use `createLogtapeSink` as the canonical factory name.
-  Refresh the LogTape forwarding migration note for consumers moving off the retired logging package.
-
-### Patch Changes
-
-- 6300f70: Refresh source comments and test labels for retired connector terminology as adapter guardrails become strict.
-- a8997ed: Add migration guidance for the retired `@ontrails/logging` package and align observability README examples around `@ontrails/observe`, `@ontrails/tracing`, and `@ontrails/logtape`.
-- d40430d: Remove the retired `@ontrails/logging` workspace from the prerelease package set. Use `@ontrails/observe` for log and trace sink contracts and `@ontrails/logtape` for LogTape forwarding.
-- Updated dependencies [6300f70]
-- Updated dependencies [e898cc4]
-- Updated dependencies [a8997ed]
-- Updated dependencies [fe03945]
-- Updated dependencies [d40430d]
-- Updated dependencies [49c2e7d]
-- Updated dependencies [9cdb0f2]
-- Updated dependencies [22c6c06]
-  - @ontrails/observe@1.0.0-beta.16
-
-## 1.0.0-beta.15
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.15
