--- packages/logtape/tsconfig.tests.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "noEmit": true,
-    "rootDir": "./src",
-    "types": ["bun"]
-  },
-  "include": ["src/**/*.test.ts", "src/__tests__/**/*.ts"],
-  "exclude": []
-}
