+++ packages/observe/README.md
@@ -12,7 +12,7 @@ Use `@ontrails/observe` for app-facing observability contracts and sinks: `LogSi
 
 `@ontrails/tracing` remains a compatibility and developer-state package for tracing-specific local tooling: query/status trails, the SQLite dev store, sampling helpers, and the supported `@ontrails/tracing/otel` OpenTelemetry adapter subpath.
 
-For v1, OpenTelemetry trace export lives at `@ontrails/tracing/otel`; there is no standalone `@ontrails/otel` package. That adapter translates Trails-native `TraceRecord` values to callback-delivered OTel-shaped spans without requiring the OpenTelemetry SDK as a runtime dependency. Use `@ontrails/pino` separately when forwarding log records to a Pino-shaped logger.
+For v1, OpenTelemetry trace export lives at `@ontrails/tracing/otel`; there is no standalone `@ontrails/otel` package. That adapter translates Trails-native `TraceRecord` values to callback-delivered OTel-shaped spans without requiring the OpenTelemetry SDK as a runtime dependency. Use `@ontrails/observe/pino` separately when forwarding log records to a Pino-shaped logger.
 
 ```typescript
 import {
@@ -31,6 +31,40 @@ const sink = combine(
 
 `createFileSink()` is append-only and does not rotate files. Use external log rotation or a production adapter when retention policy matters.
 
+## Log Forwarding Subpaths
+
+Use `@ontrails/observe/logtape` when forwarding Trails log records to an existing LogTape-shaped logger:
+
+```typescript
+import { createLogtapeSink } from '@ontrails/observe/logtape';
+
+const logger = {
+  debug(message: string, props?: Record<string, unknown>) {},
+  error(message: string, props?: Record<string, unknown>) {},
+  fatal(message: string, props?: Record<string, unknown>) {},
+  info(message: string, props?: Record<string, unknown>) {},
+  trace(message: string, props?: Record<string, unknown>) {},
+  warn(message: string, props?: Record<string, unknown>) {},
+};
+const sink = createLogtapeSink({ logger });
+```
+
+Use `@ontrails/observe/pino` when an app already owns a Pino logger:
+
+```typescript
+import { createPinoSink } from '@ontrails/observe/pino';
+
+const logger = {
+  debug(payload: Record<string, unknown>, message?: string) {},
+  error(payload: Record<string, unknown>, message?: string) {},
+  fatal(payload: Record<string, unknown>, message?: string) {},
+  info(payload: Record<string, unknown>, message?: string) {},
+  trace(payload: Record<string, unknown>, message?: string) {},
+  warn(payload: Record<string, unknown>, message?: string) {},
+};
+const sink = createPinoSink(logger);
+```
+
 ## Migration from `@ontrails/logging`
 
-`@ontrails/logging` was retired before v1. Move sink contracts, console/file sinks, formatters, and bounded memory sinks to `@ontrails/observe`. Use `@ontrails/logtape` for LogTape forwarding, `@ontrails/pino` for Pino forwarding, and `@ontrails/tracing` for tracing registry, dev-store, query/status, sampling, and OTel adapter APIs.
+`@ontrails/logging` was retired before v1. Move sink contracts, console/file sinks, formatters, and bounded memory sinks to `@ontrails/observe`. Use `@ontrails/observe/logtape` for LogTape forwarding, `@ontrails/observe/pino` for Pino forwarding, and `@ontrails/tracing` for tracing registry, dev-store, query/status, sampling, and OTel adapter APIs.
