--- packages/pino/CHANGELOG.md
@@ -1,122 +1,0 @@
-# @ontrails/pino
-
-## 1.0.0-beta.39
-
-## 1.0.0-beta.38
-
-## 1.0.0-beta.37
-
-## 1.0.0-beta.36
-
-## 1.0.0-beta.35
-
-## 1.0.0-beta.34
-
-## 1.0.0-beta.33
-
-## 1.0.0-beta.32
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.32
-
-## 1.0.0-beta.31
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.31
-
-## 1.0.0-beta.30
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.30
-
-## 1.0.0-beta.29
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.29
-
-## 1.0.0-beta.28
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.28
-
-## 1.0.0-beta.27
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.27
-
-## 1.0.0-beta.26
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.26
-
-## 1.0.0-beta.25
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.25
-
-## 1.0.0-beta.24
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.24
-
-## 1.0.0-beta.23
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.23
-
-## 1.0.0-beta.22
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.22
-
-## 1.0.0-beta.21
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.21
-
-## 1.0.0-beta.20
-
-### Patch Changes
-
-- @ontrails/observe@1.0.0-beta.20
-
-## 1.0.0-beta.19
-
-### Patch Changes
-
-- ed5926b: Add missing TSDoc for public adapter and sink boundary types.
-- Updated dependencies [1eb5bdc]
-  - @ontrails/observe@1.0.0-beta.19
-
-## 1.0.0-beta.18
-
-### Minor Changes
-
-- e504e67: Add the publishable `@ontrails/pino` package scaffold.
-- 1a65022: Implement the structural Pino log sink.
-
-### Patch Changes
-
-- bf44972: Document Pino sink usage and publish-readiness checks.
-- Updated dependencies [bf44972]
-- Updated dependencies [e0ae995]
-  - @ontrails/observe@1.0.0-beta.18
-
-## 1.0.0-beta.17
-
-### Minor Changes
-
-- Initial publishable package scaffold.
-- Add structural Pino log sink.
