--- packages/pino/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@ontrails/pino",
-  "version": "1.0.0-beta.39",
-  "files": [
-    "src/**/*.ts",
-    "!src/**/__tests__/**",
-    "!src/**/*.test.ts",
-    "!src/**/*.test-d.ts",
-    "README.md",
-    "CHANGELOG.md"
-  ],
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts",
-    "./package.json": "./package.json"
-  },
-  "scripts": {
-    "build": "tsc -b",
-    "test": "bun test",
-    "typecheck": "tsc --noEmit",
-    "lint": "oxlint ./src",
-    "clean": "rm -rf dist *.tsbuildinfo"
-  },
-  "peerDependencies": {
-    "@ontrails/observe": "workspace:^"
-  }
-}
