+++ scripts/vocab-cutover-map.ts
@@ -42,12 +42,10 @@ export const changelogHistoryPaths = [
   'packages/core/CHANGELOG.md',
   'packages/http/CHANGELOG.md',
   'packages/library/CHANGELOG.md',
-  'packages/logtape/CHANGELOG.md',
   'packages/mcp/CHANGELOG.md',
   'packages/observe/CHANGELOG.md',
   'packages/oxlint-plugin/CHANGELOG.md',
   'packages/permits/CHANGELOG.md',
-  'packages/pino/CHANGELOG.md',
   'packages/regrade/CHANGELOG.md',
   'packages/source/CHANGELOG.md',
   'packages/store/CHANGELOG.md',
